Code refactoring
Define and throw a dedicated exception instead of using a generic one.
Add curly braces around the nested statement(s).
Add a "case default" clause to this "switch" statement.

verifyCert: This method has 24 returns, which is more than the 3 allowed.
verifyCert: Refactor this function to reduce its Cognitive Complexity from 49 to the 15 allowed.